Replace styled-components to scss modules, change full page loader
