+++ src/components/pokemon-item/pokemon-item.module.scss
@@ -1,0 +1,18 @@
+.link {
+  width: 96px;
+  display: inline-block;
+  color: inherit;
+  text-decoration: none;
+}
+
+.imageWrapper {
+  display: flex;
+  align-items: center;
+  width: 96px;
+  height: 96px;
+}
+
+.cardImage {
+  width: 96px;
+  aspect-ratio: 1;
+}
